UICBACK1/apps/order/api_endpoints/OrderItemCreate/views.py
from rest_framework.generics import CreateAPIView
from .serializers import OrderItemCreateSerializer

class OrderItemCreateAPIView(CreateAPIView):
    serializer_class = OrderItemCreateSerializer
    
    