UICBACK1/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-debug-toolbar==5.2.0
django-filter==25.1
django-query-counter==0.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
Faker==37.4.0
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tabulate==0.9.0
tzdata==2025.2
uritemplate==4.2.0
